requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
aiohttp==3.9.1
aiodns==3.1.1
dnspython==2.4.2
email-validator==2.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-redis==5.4.0
django-cors-headers==4.3.1
